fix: Favicon generation
Fixed issues with favicon generation relying on the public folder. The icon will now be generated on the fly
