--- apps/portal/public/placeholder.txt
@@ -1,1 +1,0 @@
-just a placeholder to be able to publish this folder
